requirements.txt
streamlit
snowflake
snowflake-ml-python
snowflake-core
snowflake-snowpark-python
snowflake-connector-python
python-dotenv
